Sends custom user agent with cloudflare-rs
